Split color module in a folder, add color module with RGB888, RGB565, and palette support; implement color manipulation functions
